split info and select tools
