+++ tools/data_manipulation/xarray/xarray_metadata_info.xml
@@ -7,13 +7,11 @@
     <requirements>
         <requirement type="package" version="@TOOL_VERSION@">xarray</requirement>
         <requirement type="package" version="3.10">python</requirement>
-        <requirement type="package" version="1.8.2">shapely</requirement>
         <requirement type="package" version="1.6.0">netcdf4</requirement>
         <requirement type="package" version="1.4.3">pandas</requirement>
-        <requirement type="package" version="0.11.0">geopandas</requirement>
     </requirements>
     <command detect_errors="exit_code"><![CDATA[
-        python3 '$__tool_directory__/xarray_tool.py' '$input' --info '$info' --summary '$output'
+        python3 '$__tool_directory__/xarray_info.py' '$input' --info '$info' --summary '$output'
     ]]>    </command>
     <inputs>
         <param type="data" name="input" label="Netcdf file" format="netcdf,h5" help="Netcdf file you need information about."/>
